Personality: refactor
